spirv: Create a InjectFunctionPass subclass
