+++ layers/gpu/spirv/CMakeLists.txt
@@ -34,6 +34,8 @@ target_sources(gpu_av_spirv PRIVATE
     module.cpp
     type_manager.h
     type_manager.cpp
+    inject_function_pass.h
+    inject_function_pass.cpp
     pass.h
     pass.cpp
     ${VVL_SOURCE_DIR}/layers/${API_TYPE}/generated/spirv_grammar_helper.cpp
