+++ layers/gpu/spirv/bindless_descriptor_pass.h
@@ -15,22 +15,18 @@
 #pragma once
 
 #include <stdint.h>
-#include "pass.h"
+#include "inject_function_pass.h"
 
 namespace gpuav {
 namespace spirv {
-
-class Module;
-struct Function;
-struct BasicBlock;
 
 // Create a pass to instrument bindless descriptor checking
 // This pass instruments all bindless references to check that descriptor
 // array indices are inbounds, and if the descriptor indexing extension is
 // enabled, that the descriptor has been initialized.
-class BindlessDescriptorPass : public Pass {
+class BindlessDescriptorPass : public InjectFunctionPass {
   public:
-    BindlessDescriptorPass(Module& module) : Pass(module, true) {}
+    BindlessDescriptorPass(Module& module) : InjectFunctionPass(module, true) {}
 
   private:
     bool AnalyzeInstruction(const Function& function, const Instruction& inst) final;
